refactor: split XLA.jl into multiple files
